Move App component up to src
