+++ src/main.tsx
@@ -1,8 +1,8 @@
 import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
 import "./style.css";
-import App from "./components/App";
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
+import App from "./App";
 
 const queryClient = new QueryClient();
